+++ types/rest/rest.go
@@ -264,8 +264,8 @@ func PostProcessResponseBare(w http.ResponseWriter, ctx context.CLIContext, body
 	// ref: https://github.com/cosmos/cosmos-sdk/issues/5864
 	var marshaler codec.JSONMarshaler
 
-	if ctx.Marshaler != nil {
-		marshaler = ctx.Marshaler
+	if ctx.JSONMarshaler != nil {
+		marshaler = ctx.JSONMarshaler
 	} else {
 		marshaler = ctx.Codec
 	}
@@ -308,8 +308,8 @@ func PostProcessResponse(w http.ResponseWriter, ctx context.CLIContext, resp int
 	// ref: https://github.com/cosmos/cosmos-sdk/issues/5864
 	var marshaler codec.JSONMarshaler
 
-	if ctx.Marshaler != nil {
-		marshaler = ctx.Marshaler
+	if ctx.JSONMarshaler != nil {
+		marshaler = ctx.JSONMarshaler
 	} else {
 		marshaler = ctx.Codec
 	}